# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: clientMORE backend API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get tenant settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update tenant settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsPatch"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
  /files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List knowledge-base documents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentFile"
  /upload:
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload a document (parse, chunk, embed)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
  /files/{id}:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a document and its chunks
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /handoffs:
    get:
      operationId: listHandoffs
      tags: [handoffs]
      summary: List handoffs (optionally filter by status / channel)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, resolved]
        - in: query
          name: channel
          schema:
            type: string
      responses:
        "200":
          description: Handoff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Handoff"
  /handoffs/{id}/reply:
    post:
      operationId: replyHandoff
      tags: [handoffs]
      summary: Send an agent reply to the customer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required: [content]
      responses:
        "200":
          description: Updated handoff
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Handoff"
  /handoffs/{id}/resolve:
    post:
      operationId: resolveHandoff
      tags: [handoffs]
      summary: Mark a handoff as resolved
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /handoffs/{id}/feedback:
    post:
      operationId: handoffFeedback
      tags: [handoffs]
      summary: Submit CSAT score (1-5) for a resolved handoff
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                score:
                  type: integer
                  minimum: 1
                  maximum: 5
              required: [score]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /chat:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a chat message and get an AI answer (public — no auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
  /chat/{sessionId}/agent-messages:
    get:
      operationId: getAgentMessages
      tags: [chat]
      summary: Poll for agent replies (widget fallback when WebSocket is unavailable)
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
        - in: query
          name: tenant_key
          schema:
            type: string
        - in: query
          name: after_id
          schema:
            type: integer
      responses:
        "200":
          description: Agent messages since after_id
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    Settings:
      type: object
      properties:
        companyName:
          type: string
        botName:
          type: string
        companyLogo:
          type: string
        botTone:
          type: string
        systemPromptExtra:
          type: string
        telegramToken:
          type: string
          nullable: true
        isTelegramActive:
          type: boolean
        twilioSid:
          type: string
          nullable: true
        twilioToken:
          type: string
          nullable: true
        twilioNumber:
          type: string
          nullable: true
        isWhatsappActive:
          type: boolean
        subscriptionPlan:
          type: string
        usedMessages:
          type: integer
        confidenceThreshold:
          type: number
      required: [companyName, botName, botTone, isTelegramActive, isWhatsappActive, subscriptionPlan, usedMessages, confidenceThreshold]
    SettingsPatch:
      type: object
      additionalProperties: true
    DocumentFile:
      type: object
      properties:
        id:
          type: integer
        tenant_key:
          type: string
        name:
          type: string
        size:
          type: string
        type:
          type: string
        chunks:
          type: integer
        date:
          type: string
          format: date-time
        status:
          type: string
      required: [id, name, size, type, chunks, date, status]
    UploadResult:
      type: object
      properties:
        file:
          type: string
        chunks:
          type: integer
        status:
          type: string
      required: [file, chunks, status]
    Handoff:
      type: object
      properties:
        id:
          type: integer
        user:
          type: string
        channel:
          type: string
        reason:
          type: string
        language:
          type: string
        timeAgo:
          type: string
        createdAt:
          type: string
          format: date-time
        unreplied:
          type: boolean
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        metadata:
          type: object
      required: [id, user, channel, reason, timeAgo, createdAt, unreplied, messages]
    Message:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
        content:
          type: string
        time:
          type: string
          format: date-time
      required: [id, role, content, time]
...
    ChatRequest:
      type: object
      properties:
        session_id:
          type: string
        message:
          type: string
        channel:
          type: string
        tenantKey:
          type: string
        tenant_key:
          type: string
      required: [session_id, message]
    ChatResponse:
      type: object
      properties:
        answer:
          type: string
        confidence:
          type: number
        handoff:
          type: object
          nullable: true
          properties:
            id:
              type: integer
        language:
          type: string
        escalate:
          type: boolean
        sender:
          type: string
      required: [answer, confidence]
...